Fix cross-domain request in IE 8
IE8 requires use of a different object to provide XHR cross-domain.
jQuery doesn't include this object by default because including it
causes unpleasantness for other versions of IE and requires developer
knowledge to use and debug correctly.  This patches in the required
object as an ajaxTransport for jQuery only when the XDomainRequest
object is available (only when running IE8).